Refactor heavily
The functionality in this crate is awesome, but I wanted to bring it
into line with the code quality of other Rust crates. I also implemented
some of my own opinions on naming, I hope this does not offend the
author. I thank the author muchly for what they have built.